Create seperate search component
Import it in RestaurantsScreen.js, remove unwanted code and importations from the file.
